worker/requirements.txt
httpx
python-dotenv
orjson
